# storage/api-docs/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Support Plateforme API",
    "version": "1.0.0",
    "description": "API de gestion des tickets de support avec authentification Sanctum"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Serveur local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token Bearer à utiliser dans l'en-tête Authorization"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["user", "technicien", "administrateur"] },
          "actif": { "type": "boolean" }
        }
      },
...
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Connexion",
        "description": "Se connecter et recevoir un token Bearer.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token JWT retourné",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "token_type": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Identifiants invalides" }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Mot de passe oublié",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "email": { "type": "string", "format": "email" } },
                "required": ["email"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Email de réinitialisation envoyé" }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Réinitialiser le mot de passe",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "token": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Mot de passe réinitialisé" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Déconnexion",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Déconnexion réussie" },
          "401": { "description": "Non authentifié" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Profil connecté",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Utilisateur connecté",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": { "description": "Non authentifié" }
        }
      }
    },
    "/auth/profil": {
      "patch": {
        "tags": ["Auth"],
        "summary": "Mettre à jour le profil",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "current_password": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Profil mis à jour" }
        }
      }
    },
...
    "/articles/{id}": {
      "get": {
        "tags": ["Articles"],
        "summary": "Article détaillé",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Détails de l'article" },
          "404": { "description": "Article non trouvé" }
        }
      }
    },
...
  }
}